Rename TypeClassesList.h to type-classes-list.h
---
