+++ compiler/dyno/include/chpl/types/Type.h
@@ -128,7 +128,7 @@ class Type {
   virtual void stringify(std::ostream& ss, chpl::StringifyKind stringKind) const;
 
   // define is__ methods for the various Type subclasses
-  // using macros and TypeClassesList.h
+  // using macros and type-classes-list.h
   /// \cond DO_NOT_DOCUMENT
   #define TYPE_IS(NAME) \
     bool is##NAME() const { \
@@ -139,8 +139,8 @@ class Type {
   #define TYPE_BEGIN_SUBCLASSES(NAME) TYPE_IS(NAME)
   #define TYPE_END_SUBCLASSES(NAME)
   /// \endcond
-  // Apply the above macros to TypeClassesList.h
-  #include "chpl/types/TypeClassesList.h"
+  // Apply the above macros to type-classes-list.h
+  #include "chpl/types/type-classes-list.h"
   // clear the macros
   #undef TYPE_NODE
   #undef BUILTIN_TYPE_NODE
@@ -199,7 +199,7 @@ class Type {
   const CompositeType* getCompositeType() const;
 
   // define to__ methods for the various Type subclasses
-  // using macros and TypeClassesList.h
+  // using macros and type-classes-list.h
   // Note: these offer equivalent functionality to C++ dynamic_cast<DstType*>
   /// \cond DO_NOT_DOCUMENT
   #define TYPE_TO(NAME) \
@@ -214,8 +214,8 @@ class Type {
   #define TYPE_BEGIN_SUBCLASSES(NAME) TYPE_TO(NAME)
   #define TYPE_END_SUBCLASSES(NAME)
   /// \endcond
-  // Apply the above macros to TypeClassesList.h
-  #include "chpl/types/TypeClassesList.h"
+  // Apply the above macros to type-classes-list.h
+  #include "chpl/types/type-classes-list.h"
   // clear the macros
   #undef TYPE_NODE
   #undef BUILTIN_TYPE_NODE
