+++ compiler/dyno/include/chpl/types/TypeClasses.h
@@ -25,7 +25,7 @@ namespace types {
 
 
 // forward declare the various Type subclasses
-// using macros and TypeClassesList.h
+// using macros and type-classes-list.h
 /// \cond DO_NOT_DOCUMENT
 #define TYPE_DECL(NAME) class NAME;
 #define TYPE_NODE(NAME) TYPE_DECL(NAME)
@@ -33,8 +33,8 @@ namespace types {
 #define TYPE_BEGIN_SUBCLASSES(NAME) TYPE_DECL(NAME)
 #define TYPE_END_SUBCLASSES(NAME)
 /// \endcond
-// Apply the above macros to TypeClassesList.h
-#include "chpl/types/TypeClassesList.h"
+// Apply the above macros to type-classes-list.h
+#include "chpl/types/type-classes-list.h"
 // clear the macros
 #undef TYPE_NODE
 #undef BUILTIN_TYPE_NODE
